Refactor tasks and patch services out of main
