+++ app/tasks/__init__.py
@@ -1,0 +1,33 @@
+from .service import (
+    PATCH_APPLIED_DIR_REL,
+    PATCH_PROPOSALS_DIR_REL,
+    RUN_CHECKS_DIR_REL,
+    TASKS_DONE_DIR_REL,
+    TASKS_OPEN_DIR_REL,
+    code_checks_run_service,
+    code_merge_service,
+    code_patch_propose_service,
+    docs_patch_apply_service,
+    docs_patch_propose_service,
+    load_check_artifacts,
+    tasks_create_service,
+    tasks_query_service,
+    tasks_update_service,
+)
+
+__all__ = [
+    "PATCH_APPLIED_DIR_REL",
+    "PATCH_PROPOSALS_DIR_REL",
+    "RUN_CHECKS_DIR_REL",
+    "TASKS_DONE_DIR_REL",
+    "TASKS_OPEN_DIR_REL",
+    "code_checks_run_service",
+    "code_merge_service",
+    "code_patch_propose_service",
+    "docs_patch_apply_service",
+    "docs_patch_propose_service",
+    "load_check_artifacts",
+    "tasks_create_service",
+    "tasks_query_service",
+    "tasks_update_service",
+]
